[REFACTOR] Move token consensus functionality to common libs
