+++ src/Makefile.am
@@ -172,6 +172,7 @@ BITCOIN_CORE_H = \
   tinyformat.h \
   tokens/tokendb.h \
   tokens/tokengroupconfiguration.h \
+  tokens/tokengroupdescription.h \
   tokens/tokengroupmanager.h \
   torcontrol.h \
   txdb.h \
@@ -226,7 +227,6 @@ libbitcoin_server_a_SOURCES = \
   blocksignature.cpp \
   chain.cpp \
   checkpoints.cpp \
-  consensus/tokengroups.cpp \
   httprpc.cpp \
   httpserver.cpp \
   init.cpp \
@@ -447,6 +447,7 @@ libbitcoin_common_a_SOURCES = \
   chainparams.cpp \
   coins.cpp \
   compressor.cpp \
+  consensus/tokengroups.cpp \
   dstencode.cpp \
   ionaddrenc.cpp \
   primitives/block.cpp \
@@ -470,6 +471,7 @@ libbitcoin_common_a_SOURCES = \
   script/script_error.cpp \
   spork.cpp \
   sporkdb.cpp \
+  tokens/tokengroupdescription.cpp \
   $(BITCOIN_CORE_H)
 
 # util: shared between all executables.
