+++ src/tokens/tokengroupmanager.h
@@ -39,9 +39,6 @@ public:
     bool GetTokenGroupIdByName(std::string strName, CTokenGroupID &tokenGroupID);
     std::map<CTokenGroupID, CTokenGroupCreation> GetMapTokenGroups() { return mapTokenGroups; };
 
-    bool BuildGroupDescData(CScript script, std::vector<std::vector<unsigned char> > &descriptionData);
-    bool ParseGroupDescData(const CTokenGroupInfo &tgInfo, const std::vector<std::vector<unsigned char> > descriptionData, CTokenGroupDescription &tokenGroupDescription);
-
     bool StoreManagementTokenGroups(CTokenGroupCreation tokenGroupCreation);
     void ClearManagementTokenGroups();
 
@@ -53,14 +50,13 @@ public:
     CTokenGroupID GetDarkMatterID() { return tgDarkMatterCreation->tokenGroupInfo.associatedGroup; };
     CTokenGroupID GetAtomID() { return tgAtomCreation->tokenGroupInfo.associatedGroup; };
 
-    bool MagicTokensCreated();
-    bool DarkMatterTokensCreated();
-    bool AtomTokensCreated();
+    bool MagicTokensCreated() { return tgMagicCreation ? true : false; };
+    bool DarkMatterTokensCreated() { return tgDarkMatterCreation ? true : false; };
+    bool AtomTokensCreated() { return tgAtomCreation ? true : false; };
 
     bool ManagementTokensCreated() {
         return MagicTokensCreated() && DarkMatterTokensCreated() && AtomTokensCreated();
     }
-    bool IsManagementTokenInput(CScript script);
 
     unsigned int GetTokenTxStats(const CTransaction &tx, const CCoinsViewCache& view, const CTokenGroupID &tgId, unsigned int &nTokenCount, CAmount &nTokenMint);
